add group history
