+++ src/app/(authed)/groups/page.tsx
@@ -1,4 +1,4 @@
-import { GroupsWrapper } from "~/app/(authed)/groups/components/GroupsWrapper";
+import { GroupsWrapper } from "~/app/(authed)/groups/_components/GroupsWrapper";
 
 export default function GroupPage() {
   /** @ts-expect-error Async Component  */
